Эксперимент с аналайзером, теперь модуль работает примено то время, которое необходимо.

+++ app/image_processing_2d_test_app/src/data/pixelFrequencies.h
@@ -15,7 +15,7 @@ struct PixelFrequencies
 
 	unsigned int _counter;
 
-    unsigned long _pixel0UnassignedFrequency;
+    unsigned int _pixel0UnassignedFrequency;
 };
